Admin role added, admin can add beer to DB

+++ Kortyérzet/Domain/Beer.cs
@@ -12,14 +12,14 @@ namespace Kortyérzet.Domain
         public string Logo { get; set; }
         public string Style { get; set; }
         public string Description { get; set; }
-        public int Rating { get; set; }
+        public float Rating { get; set; }
         public int TimesRated { get; set; }
         public int CheckIn { get; set; }
         public float ABV { get; set; }
         public int IBU { get; set; }
         public int BreweryID { get; set; }
 
-        public Beer(int id, string name, string logo, string style, string description, int rating, int timesRated, int checkIn, float abv, int ibu, int breweryID)
+        public Beer(int id, string name, string logo, string style, string description, float rating, int timesRated, int checkIn, float abv, int ibu, int breweryID)
         {
             ID = id;
             Name = name;
